refactor(frontend): move ic transaction types to own module
