+++ src/frontend/src/icp/components/transactions/IcTransactions.svelte
@@ -21,7 +21,7 @@
 	} from '$icp/derived/ic-token.derived';
 	import { icTransactions } from '$icp/derived/ic-transactions.derived';
 	import { loadNextTransactions } from '$icp/services/ic-transactions.services';
-	import type { IcTransactionUi } from '$icp/types/ic';
+	import type { IcTransactionUi } from '$icp/types/ic-transaction';
 	import TransactionsPlaceholder from '$lib/components/transactions/TransactionsPlaceholder.svelte';
 	import Header from '$lib/components/ui/Header.svelte';
 	import { WALLET_PAGINATION } from '$lib/constants/app.constants';
