+++ src/frontend/src/icp/schedulers/ic-wallet.scheduler.ts
@@ -1,4 +1,4 @@
-import type { IcTransactionAddOnsInfo, IcTransactionUi } from '$icp/types/ic';
+import type { IcTransactionAddOnsInfo, IcTransactionUi } from '$icp/types/ic-transaction';
 import type { GetTransactions } from '$icp/types/ic.post-message';
 import { queryAndUpdate } from '$lib/actors/query.ic';
 import { WALLET_TIMER_INTERVAL_MILLIS } from '$lib/constants/app.constants';
