+++ src/frontend/src/icp/stores/ic-pending-transactions.store.ts
@@ -1,5 +1,5 @@
 import { type IcCertifiedTransaction } from '$icp/stores/ic-transactions.store';
-import type { IcTransactionUi } from '$icp/types/ic';
+import type { IcTransactionUi } from '$icp/types/ic-transaction';
 import { initCertifiedStore, type CertifiedStore } from '$lib/stores/certified.store';
 import type { CertifiedData } from '$lib/types/store';
 import type { TokenId } from '$lib/types/token';
