+++ src/frontend/src/icp/types/btc.ts
@@ -1,4 +1,4 @@
-import type { IcTransactionIdText } from '$icp/types/ic';
+import type { IcTransactionIdText } from '$icp/types/ic-transaction';
 import type { RetrieveBtcStatusV2 } from '@dfinity/ckbtc';
 
 export type BtcWithdrawalStatuses = Record<IcTransactionIdText, RetrieveBtcStatusV2>;
