+++ src/frontend/src/icp/utils/ic-transactions.utils.ts
@@ -1,14 +1,13 @@
 import { ICP_TOKEN_ID } from '$env/tokens.env';
 import type { BtcStatusesData } from '$icp/stores/btc.store';
 import type { IcCertifiedTransaction } from '$icp/stores/ic-transactions.store';
+import type { IcCkToken, IcToken } from '$icp/types/ic';
 import type {
-	IcCkToken,
-	IcpTransaction,
-	IcrcTransaction,
-	IcToken,
 	IcTransaction,
-	IcTransactionUi
-} from '$icp/types/ic';
+	IcTransactionUi,
+	IcpTransaction,
+	IcrcTransaction
+} from '$icp/types/ic-transaction';
 import { extendCkBTCTransaction, mapCkBTCTransaction } from '$icp/utils/ckbtc-transactions.utils';
 import { mapCkEthereumTransaction } from '$icp/utils/cketh-transactions.utils';
 import {
