+++ src/frontend/src/tests/mocks/ckbtc.mock.ts
@@ -1,5 +1,5 @@
 import { BTC_MAINNET_EXPLORER_URL } from '$env/explorers.env';
-import type { IcTransactionUi } from '$icp/types/ic';
+import type { IcTransactionUi } from '$icp/types/ic-transaction';
 import { utxoTxIdToString } from '$icp/utils/btc.utils';
 import type { MinterInfo, PendingUtxo } from '@dfinity/ckbtc';
